GLPG-GL: Fix context creation on Intel
For some reason, GL context creation was failing
on Intel Gen iGPUs since calls to SetPixelFormat
for creating the actual GL context via WGL were returning
a failure.

This was because SetPixelFormat was being called twice
on the same hDc, once for creating the dummy GL context
and a second time when creating the actual GL context
and as per MSDN, SetPixelFormat can only
be called once on a hDc.

This CL tweaks the WGL parts of the codebase by
creating a temporary GLPGWindow and using its hDc
to create the temporary GL context